added more foundation for role based authentication

+++ src/main/java/com/plumauto/service/AutomationEngine.java
@@ -3,11 +3,15 @@ package com.plumauto.service;
 import com.plumauto.entity.BuildDetails;
 import com.plumauto.entity.JobDetail;
 import com.plumauto.entity.RunDetails;
+import com.plumauto.entity.UserDetails;
 import com.plumauto.repository.Job;
+import com.plumauto.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.io.*;
@@ -26,23 +30,47 @@ public class     AutomationEngine {
     @Autowired
     JobScanner runner;
 
+    @Autowired
+    UserRepository userRepository;
+
     private final Path rootPath;
     public AutomationEngine(@Value("${app.auto-engine.path}") String rootPath) {
         this.rootPath = Paths.get(rootPath);
 
     }
 
-    public boolean createJob(JobDetail jobDetail) throws IOException, InterruptedException {
+    public boolean isJobAccessAvailable(JobDetail jobDetail){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        assert auth != null;
+        String user = auth.getName();
+        UserDetails userDetails = userRepository.findByUsername(user);
+        return jobDetail.getJobConfig().getAllowedUsers().contains(user) &&
+                userDetails.getOrganisationName() != null &&
+                jobDetail.getJobConfig().getAllowedOrganization().equals(userDetails.getOrganisationName());
+    }
+
+    public boolean createJob(JobDetail jobDetail) throws Exception {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        assert auth != null;
+        String user = auth.getName();
+        UserDetails userDetails = userRepository.findByUsername(user);
+
+        if (userRepository.findByUsername(user).getOrganisationName() == null) {
+            throw new Exception("No Org joined");
+        }
 
         List<String> validateUsername = List.of(jobDetail.getJobName().split(" "));
         if (validateUsername.size() > 1) {
             throw new RuntimeException("Invalid Job Name: Job name should not contain spaces");
         }
+
         if (jobRepository.findByJobName(jobDetail.getJobName()) != null) {
             log.error("Job already exists with name: " + jobDetail.getJobName());
             throw new RuntimeException("Job already exists with name: " + jobDetail.getJobName());
         }
-        Files.createDirectories(Path.of(rootPath + "/" + jobDetail.getJobName()));
+
+        Files.createDirectories(Path.of(rootPath + "/" + userDetails.getOrganisationName() +"/"+jobDetail.getJobName()));
+
         List<BuildDetails> tempBuildInfo = new ArrayList<>();
         BuildDetails buildInfo = new BuildDetails();
         buildInfo.setBuildNumber("0");
@@ -50,18 +78,27 @@ public class     AutomationEngine {
         buildInfo.setCreatedAt(LocalDateTime.now());
         buildInfo.setStatus("completed");
         tempBuildInfo.add(buildInfo);
+
+        if(jobDetail.getJobConfig().getAllowedUsers()==null){
+            jobDetail.getJobConfig().setAllowedUsers(new ArrayList<>());
+        }
         jobDetail.setBuildNumber(tempBuildInfo);
+        jobDetail.getJobConfig().getAllowedUsers().add(userDetails.getUsername());
+        jobDetail.getJobConfig().setAllowedOrganization(userDetails.getOrganisationName());
         jobRepository.save(jobDetail);
         return true;
     }
 
 
-    public boolean deleteJob(String jobName) throws IOException, InterruptedException {
+    public boolean deleteJob(String jobName) throws Exception {
         JobDetail job = jobRepository.findByJobName(jobName);
         if (job == null) {
             throw new RuntimeException("Job not found: " + jobName);
         } else {
-            FileUtils.deleteDirectory(new File(rootPath + "/" + job.getJobName()));
+            if(!isJobAccessAvailable(job)){
+                throw new Exception("You do not have access to this job");
+            }
+            FileUtils.deleteDirectory(new File(rootPath + "/" + job.getJobConfig().getAllowedOrganization() +"/"+job.getJobName()));
             jobRepository.delete(job);
             return true;
         }
@@ -75,8 +112,8 @@ public class     AutomationEngine {
         } else if (jobDetail.getJobName().split(" ").length > 1) {
             throw new RuntimeException("Invalid Job Name: Job name should not contain spaces");
         } else {
-            File oldDirName = new File(rootPath + "/" + job.get().getJobName());
-            File newDirName = new File(rootPath + "/" + jobDetail.getJobName());
+            File oldDirName = new File(rootPath + "/" + job.get().getJobConfig().getAllowedOrganization() +"/"+job.get().getJobConfig());
+            File newDirName = new File(rootPath + "/" + jobDetail.getJobConfig().getAllowedOrganization() +"/"+jobDetail.getJobName());
             if (oldDirName.renameTo(newDirName)) {
                 job.get().setJobName(jobDetail.getJobName().trim());
                 job.get().setDescription(jobDetail.getDescription());
@@ -96,17 +133,19 @@ public class     AutomationEngine {
         }
         String BuildNumber = String.valueOf(job.getBuildNumber().size());
 
-        RunDetails runDetails = new RunDetails();
-        runDetails.setJobDetail(job);
-        runDetails.setBuildNumber(BuildNumber);
-        runner.getPendingTaskQueue().add(runDetails);
-
         BuildDetails buildInfo = new BuildDetails();
         buildInfo.setBuildNumber(BuildNumber);
         buildInfo.setStatus("in-progress");
         buildInfo.setCreatedAt(LocalDateTime.now());
         job.getBuildNumber().add(buildInfo);
         jobRepository.save(job);
+
+        RunDetails runDetails = new RunDetails();
+        runDetails.setJobDetail(job);
+        runDetails.setBuildNumber(BuildNumber);
+        runner.getPendingTaskQueue().add(runDetails);
+
+
         return true;
     }
